+++ src/sbml/packages/fbc/sbml/FbcOr.cpp
@@ -276,7 +276,7 @@ FbcOr::toInfix() const
   for (size_t pos = 1; pos < mAssociations.size(); ++pos)
   {
     str << " or ";
-    str << mAssociations.get(pos)->toInfix();
+    str << mAssociations.get((unsigned int)pos)->toInfix();
   }
   str << ")";
   return str.str();
@@ -606,20 +606,20 @@ FbcOr::readAttributes (const XMLAttributes& attributes,
   if (getErrorLog() != NULL)
   {
     numErrs = getErrorLog()->getNumErrors();
-    for (int n = numErrs-1; n >= 0; n--)
+    for (int n = (int)numErrs-1; n >= 0; n--)
     {
-      if (getErrorLog()->getError(n)->getErrorId() == UnknownPackageAttribute)
+      if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownPackageAttribute)
       {
         const std::string details =
-                          getErrorLog()->getError(n)->getMessage();
+                          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownPackageAttribute);
         getErrorLog()->logPackageError("fbc", FbcUnknown,
                        getPackageVersion(), sbmlLevel, sbmlVersion, details, getLine(), getColumn());
       }
-      else if (getErrorLog()->getError(n)->getErrorId() == UnknownCoreAttribute)
+      else if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownCoreAttribute)
       {
         const std::string details =
-                          getErrorLog()->getError(n)->getMessage();
+                          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownCoreAttribute);
         getErrorLog()->logPackageError("fbc", FbcUnknown,
                        getPackageVersion(), sbmlLevel, sbmlVersion, details, getLine(), getColumn());
